requirements.txt
autoreject==0.4.3
coffeine==0.1
edfrd==0.7
h5io==0.2.4
h5py==3.7.0
imblearn==0.0
joblib==1.1.0
matplotlib
mne==1.9.0
mne_bids==0.16.0
mne_features==0.3
pandas==1.4.4
psutil==5.9.8
PyYAML==6.0.2
scikit_learn==1.0.2
scipy==1.9.1
seaborn==0.11.2
